refactor(silencer): migrate ChatPanel + collapse ScreenContext shim layer (Stage C)
ChatPanel at clients/silencer/src/ui/screens/lobby/panels/. Owns the
chat box (channel name overlay, scrollback textbox, presence textbox,
chat input, scrollbar). Build attaches to the lobby parent and sets
its activeobject + ActiveChanged; Tick drains world.lobby.chatmessages,
rebuilds presence on presencechanged, refreshes the channel overlay
(and snapshots world.lobby.lastchannel on first change), and flashes
the OS taskbar/dock on chat arrival when Game::minimized is set.

What moved out of Game:
- CreateChatInterface body → ChatPanel::Build
- TEXTBOX uid=9 (presence) + TEXTBOX else (chat scrollback) arms in
  ProcessLobbyInterface → ChatPanel::Tick
- OVERLAY uid=1 (channel name) arm in ProcessLobbyInterface →
  ChatPanel::Tick
- TEXTINPUT chat-send branch in ProcessLobbyInterface → ChatPanel::Tick
  (the enterpressed=false reset stays in ProcessLobbyInterface for
   un-migrated game-create / password form inputs)
- chatlinesprinted Game member deleted (was always 0; legacy tests
  against it become "chatmessages.empty()")
- lastchannel Game member moved onto the client-side Lobby class,
  next to the existing channel/channelchanged state (GoBack +
  TickLobbyBody + ChatPanel all read world.lobby.lastchannel)
- iface->activeobject + iface->ActiveChanged calls moved out of
  CreateLobbyInterface into ChatPanel::Build (the chat iface didn't
  exist when CreateLobbyInterface ran, so the assignment was setting
  activeobject = chat-iface-id-from-prior-frame anyway)

Architecture cleanup the user pushed back on, folded into this commit:

- ScreenContext is now a pure subsystem-refs holder + global
  state-machine actions. Per-screen / per-panel logic does not live
  there. Screens reach Game directly through the new public
  ScreenContext::game ref.

- All ad-hoc ScreenContext shims I'd added during stages A/B
  deleted: BuildLegacyLobbyInterface, TickLegacyLobbyBody,
  GetSelectedAgency, SetCharacterInterfaceId, NotifyAgencyChanged.
  Callers reach the Game members/methods directly.

- Game::GetSelectedAgency deleted entirely. The lobby pump's
  joining/create flows + JoinGame now read
  Config::GetInstance().defaultagency, which is already the source
  of truth (CharacterPanel mirrors every toggle change to it). The
  CharacterPanel itself uses a private ReadSelectedAgency helper to
  detect transitions (Config holds the post-update value, not useful
  for change-detection).

- Game::characterinterface deleted (only used by the deleted
  GetSelectedAgency).

- World::IsConnected() added as a public method (just exposes the
  state==CONNECTED test). CharacterPanel::Tick calls it directly
  rather than going through a Game shim.

- Game::lobbyinterface, Game::chatinterface, Game::minimized are
  public temp scaffolding; Game::CreateLobbyInterface and
  Game::TickLobbyBody are also public temp scaffolding. All five
  go away in stage H once the legacy helpers are deleted.

Verified: clean regular + unity Debug builds (only pre-existing
warnings). E2E 00_ping / 10_navigate / 20_screenshot pass.
Interactive chat smoke not exercised — set_text op only targets
TEXTBOX widgets (pre-existing CLI limitation), so driving the
lobby_connect login form headlessly to reach the lobby screen
isn't currently possible.
