+++ clients/silencer/src/ui/screens/lobby/lobby_screen.cpp
@@ -1,24 +1,29 @@
 #include "lobby_screen.h"
 
 #include "screen_context.h"
+#include "game.h"
 #include "world.h"
 #include "interface.h"
 #include "objecttypes.h"
 
 void LobbyScreen::Build(ScreenContext & ctx)
 {
-	interfaceId = ctx.BuildLegacyLobbyInterface();
-	Interface * lobbyiface = (Interface *)ctx.world.GetObjectFromId(interfaceId);
+	Interface * lobbyiface = ctx.game.CreateLobbyInterface();
+	ctx.game.lobbyinterface = lobbyiface->id;
+	interfaceId = lobbyiface->id;
 	character.Build(ctx, lobbyiface);
+	chat.Build(ctx, lobbyiface);
 }
 
 void LobbyScreen::Tick(ScreenContext & ctx)
 {
-	ctx.TickLegacyLobbyBody();
+	ctx.game.TickLobbyBody();
 	character.Tick(ctx);
+	chat.Tick(ctx);
 }
 
 void LobbyScreen::Destroy(ScreenContext & ctx)
 {
 	character.Destroy(ctx);
+	chat.Destroy(ctx);
 }
