+++ clients/silencer/src/world/world.cpp
@@ -1936,6 +1936,10 @@ bool World::IsAuthority(void){
 	return mode == AUTHORITY;
 }
 
+bool World::IsConnected() const {
+	return state == CONNECTED;
+}
+
 void World::Illuminate(void){
 	illuminate = GASLoader::Get().world.illuminateLevel;
 }
